Radically simplify stylesheets.
In the header, there is no longer any text. It's just the logo,
centered. There's a bar along the top. Both the bar and the logo
cycle through Solarized's rainbow of colors.

The footer is gone, as is the colophon at the end of posts. The
date of the post is nestled underneath the title now.