Add social widgets.
Added the Facebook like button, the Google +1 button, and the Twitter share button. I'm interested to see if anyone clicks on them.

They show up in the top right. If the page is narrower than 768 pixels, they'll disappear (so they don't fight with the logo). I chose that width somewhat arbitrarily, as it's the width of an iPad in portrait mode. Meaning the widgets will always show up on an iPad, but never show up on an iPhone.