Fix minor bugs with social widgets.
The placeholder for the Google +1 widget was 2 pixels too short.
The actual widget is too short, too, but I don't have any control
over that. No reason for the placeholder to be weird, though.

The reference to "element" within the widgets' onmouseover functions
was incorrect. Changed to "this" since that's what it should've
been in the first place.